[BE] refactor: Post API 도메인 분리
- issue (#112)
