+++ server/src/domains/post/dto/PostDto.ts
@@ -1,4 +1,4 @@
-import { IsDate, IsNumber, IsString, IsEnum } from 'class-validator';
+import { IsDate, IsNumber, IsString, IsEnum, MaxLength, MinLength } from 'class-validator';
 import { Post, PostType } from '../models/Post';
 
 export class PostDto {
@@ -24,8 +24,12 @@ export class PostContentDto {
   @IsNumber()
   groupId: number;
   @IsString()
+  @MinLength(1)
+  @MaxLength(100)
   title: string;
   @IsString()
+  @MinLength(1)
+  @MaxLength(1023)
   content: string;
   @IsEnum(PostType)
   type: PostType;
